+++ CHANGELOG.md
@@ -44,6 +44,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Vulnerability: Merurial Repository Found
 - Vulnerability: Flask Debug Mode
 - Vulnerability: Drupal backup_migrate
+- Vulnerability: Log4j RCE
 - GIF Favicon
 - Unauthenticated Gitlab SSRF CVE 2021-22214 Demonstration
 - Software Versions List